renamed ImageHelper to Bitmap and added comments
