+++ src/client/graphics/Texture.cpp
@@ -4,6 +4,8 @@
 #include <glad/gl.h>
 #include <stb_image.h>
 
+#include "Bitmap.h"
+
 Texture::Texture() : m_id(0), m_path(""), m_width(0), m_height(0) { }
 
 Texture::Texture(unsigned int id, const std::string& path, int width, int height) 
@@ -84,14 +86,8 @@ Texture Texture::create(const std::string& path, unsigned int type)
     return Texture(texture, path, width, height);
 }
 
-Texture Texture::create(const unsigned char *rawTexture, int width, int height, unsigned int type)
+Texture Texture::create(const Bitmap &rawTexture, unsigned int type)
 {
-    if (rawTexture == nullptr)
-    {
-        std::clog << "Failed to load texture" << std::endl;
-        return Texture(0, "", 0, 0);
-    }
-
     unsigned int texture;
 
     glCreateTextures(type, 1, &texture);
@@ -103,9 +99,9 @@ Texture Texture::create(const unsigned char *rawTexture, int width, int height, 
     glTextureParameteri(texture, GL_TEXTURE_MIN_FILTER, GL_NEAREST);
     glTextureParameteri(texture, GL_TEXTURE_MAG_FILTER, GL_NEAREST);
 
-    glTextureStorage2D(texture, 1, GL_RGBA8, width, height);
-    glTextureSubImage2D(texture, 0, 0, 0, width, height, GL_RGBA, GL_UNSIGNED_BYTE, rawTexture);
+    glTextureStorage2D(texture, 1, GL_RGBA8, rawTexture.getWidth(), rawTexture.getHeight());
+    glTextureSubImage2D(texture, 0, 0, 0, rawTexture.getWidth(), rawTexture.getHeight(), GL_RGBA, GL_UNSIGNED_BYTE, rawTexture.getRawPixels().data());
     glGenerateTextureMipmap(texture);
 
-    return Texture(texture, "no_path", width, height);
+    return Texture(texture, "no_path", rawTexture.getWidth(), rawTexture.getHeight());
 }
